before production

+++ src/analyses/views.py
@@ -6,6 +6,8 @@ from django.views.generic import TemplateView, ListView, CreateView
 from models import Analysis
 from forms import AnalysisForm
 
+from tasks import xhmm
+
 # Create your views here.
 def index(request):
 
@@ -22,20 +24,21 @@ def create(request):
     if request.method == 'POST': # If the form has been submitted...
         form = AnalysisForm(request.POST) # A form bound to the POST data
         if form.is_valid(): # All validation rules pass
-            
-            
             # analysis = form.save(commit=False)
             analysis = AnalysisForm(request.POST)
 
     
             #schedule task
-                        
+
 
             # print 'request.user', request.user
 
             # analysis.user = request.user
 
-            analysis.save()
+            analysis = analysis.save()
+
+            xhmm.delay(analysis.id)
+
 
             return redirect('analyses')
 
@@ -61,15 +64,29 @@ class AnalysisCreate(CreateView):
 def bulk_actions(request):
 
     print 'entrou no bulk'
+    # print 'actions', actions
     if request.method == 'POST': 
       
+        action = request.POST.get("actions_list", "")
+        print 'action', action
+
         analyses = request.POST.getlist('analyses')
 
-        print 'analyses', analyses
-        for analysis_id in analyses:
-            instance = Analysis.objects.get(id=analysis_id)
-            instance.delete()
 
+        if action == 'delete':
+            for analysis_id in analyses:
+                instance = Analysis.objects.get(id=analysis_id)
+                instance.delete()
+        elif action == 'rerun':
+            print 'rerun'
+            for analysis_id in analyses:
+                analysis = Analysis.objects.get(id=analysis_id)
+                print 'analysis', analysis.name
+                print 'type', analysis.type
+                xhmm.delay(analysis.id)
+                            
+        print 'analyses', analyses
+        
     #get select objects
